On activation, if our data file can not be read, deactivate the module and add the temp notice

+++ inc/modules/firewall/plugins/geoip-system.php
@@ -28,12 +28,16 @@ function secupress_geoip2country( $ip ) {
  **/
 add_action( 'secupress_activate_plugin_' . basename( __FILE__, '.php' ), 'secupress_geoip_activation' );
 function secupress_geoip_activation() {
-	global $wpdb;
+	global $wpdb, $current_user;
 
 	$filename = SECUPRESS_INC_PATH . 'data/geoips.data';
 	$queries  = file_exists( $filename ) ? file_get_contents( $filename ) : false;
 	if ( ! $queries ) {
-		return; //// desactiver le plugin + signaler en notice ?
+		secupress_add_transient_notice( sprintf( __( 'The module GeoIP Management has not been activated because the file %s can not be read.', 'secupress' ), '<code>' . str_replace( realpath( ABSPATH ), '', $filename ) . '</code>' ), 'error' );
+		secupress_manage_submodule( 'firewall', 'geoip-system', false ); // deactivate the plugin
+		delete_transient( "secupress_module_activation_{$current_user->ID}" );
+		delete_transient( "secupress_module_deactivation_{$current_user->ID}" );
+		return;
 	}
 
 	if ( $wpdb->get_var( 'SHOW TABLES LIKE "' . SECUPRESS_GEOIP_TABLE .'"' ) != SECUPRESS_GEOIP_TABLE ) {
@@ -54,6 +58,18 @@ function secupress_geoip_activation() {
 			$wpdb->query( 'INSERT INTO ' . SECUPRESS_GEOIP_TABLE . ' (begin_ip, end_ip, country_code) VALUES (' . $query . ')' );
 		}
 		update_option( 'secupress_geoip_installed', 1 );
+
+		$country_code = secupress_geoip2country( secupress_get_ip() );
+		$is_whitelist = secupress_get_module_option( 'geoip-system_type', -1, 'firewall' ) == 'whitelist';
+		$countries    = array_flip( secupress_get_module_option( 'geoip-system_countries', -1, 'firewall' ) );
+		if ( ( isset( $countries[ $country_code ] ) && ! $is_whitelist ) ||
+			( ! isset( $countries[ $country_code ] ) && $is_whitelist ) ) {
+			$countries   = array_flip( $countries );
+			$countries[] = $country_code;
+		}
+		$settings = array( 'geoip-system_countries' => $countries );
+		secupress_update_module_options( $settings, 'firewall' );
+
 	}
 }
 
@@ -67,7 +83,10 @@ function secupress_geoip_activation() {
 add_action( 'secupress_deactivate_plugin_' . basename( __FILE__, '.php' ), 'secupress_geoip_deactivation' );
 function secupress_geoip_deactivation() {
 	global $wpdb;
-	$wpdb->query( 'DROP TABLE ' . SECUPRESS_GEOIP_TABLE );
+
+	if ( $wpdb->get_var( 'SHOW TABLES LIKE "' . SECUPRESS_GEOIP_TABLE .'"' ) == SECUPRESS_GEOIP_TABLE ) {
+		$wpdb->query( 'DROP TABLE ' . SECUPRESS_GEOIP_TABLE );
+	}
 	delete_option( 'secupress_geoip_installed' );
 }
